+++ nacc_gear_form_classifier/__init__.py
@@ -1,4 +1,4 @@
-"""The fw_gear_file_classifier package."""
+"""The nacc_gear_form_classifier package."""
 import importlib.metadata
 
 PKG_NAME = __name__
