+++ nacc_gear_form_classifier/classification_profiles/main.yml
@@ -2,4 +2,3 @@
 name: JSON Form Classifier
 includes:
   - json_forms.yml
-  - MR.yaml
