+++ run.py
@@ -5,8 +5,8 @@ import typing as t
 
 from flywheel_gear_toolkit import GearToolkitContext
 
-from fw_gear_file_classifier.main import classify
-from fw_gear_file_classifier.parser import parse_config
+from nacc_gear_form_classifier.main import classify
+from nacc_gear_form_classifier.parser import parse_config
 
 log = logging.getLogger(__name__)
